snaptidy/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "snaptidy"
version = "0.1.0"
authors = [
    {name = "Yongsok Kwon", email = "saisiot.dev@gmail.com"},
]
description = "A CLI tool for organizing files and removing duplicates from directories"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pillow>=9.0.0",
    "imagehash>=4.3.0",
    "exifread>=3.0.0",
    "hachoir>=3.1.3",
    "opencv-python>=4.5.0",
    "rich>=12.0.0",
]

[project.urls]
"Homepage" = "https://github.com/saisiot/snaptidy"
"Bug Tracker" = "https://github.com/saisiot/snaptidy/issues"

[project.scripts]
snaptidy = "snaptidy.cli:main"